float: All testcases now passes. (fixed average for floats that returns float on sum())

+++ src/tightdb/table_accessors.hpp
@@ -1302,7 +1302,7 @@ public:
     double average(std::size_t* resultcount=NULL, std::size_t start=0,
                    std::size_t end=std::size_t(-1), std::size_t limit=std::size_t(-1)) const
     {
-        return Base::m_query->m_impl.average<double, float>(col_idx, resultcount, start, end, limit);
+        return Base::m_query->m_impl.average<float, float>(col_idx, resultcount, start, end, limit);
     }
 };
